Add tests for thread_update

+++ misago/migrations/135f742c6a0d_threads_posts.py
@@ -32,7 +32,7 @@ def upgrade():
         sa.Column("slug", sa.String(length=255), nullable=False),
         sa.Column("started_at", sa.DateTime(), nullable=False),
         sa.Column("last_posted_at", sa.DateTime(), nullable=False),
-        sa.Column("ordering", sa.Integer(), nullable=False),
+        sa.Column("ordering", sa.BigInteger(), nullable=False),
         sa.Column("replies", sa.Integer(), nullable=False),
         sa.Column("is_closed", sa.Boolean(), nullable=False),
         sa.Column("extra", sa.JSON(), nullable=False),
