+++ misago/tables.py
@@ -115,7 +115,7 @@ threads = sqlalchemy.Table(
     sqlalchemy.Column("slug", sqlalchemy.String(length=255), nullable=False),
     sqlalchemy.Column("started_at", sqlalchemy.DateTime, nullable=False),
     sqlalchemy.Column("last_posted_at", sqlalchemy.DateTime, nullable=False),
-    sqlalchemy.Column("ordering", sqlalchemy.Integer, nullable=False),
+    sqlalchemy.Column("ordering", sqlalchemy.BigInteger, nullable=False),
     sqlalchemy.Column("replies", sqlalchemy.Integer, nullable=False),
     sqlalchemy.Column("is_closed", sqlalchemy.Boolean, nullable=False),
     sqlalchemy.Column("extra", sqlalchemy.JSON(), nullable=False),
